feat: 1. small refactor 2. added some function for viewer 3. added classes for migration of code - first 4. now Mail the VIEWERS is mandatory 5. Did the books the COUNT of the INT type

+++ Library.Data/BOOK.cs
@@ -19,7 +19,7 @@ namespace Library.Data
         public string AUTHOR { get; set; }
         public string PUBLISHING { get; set; }
 
-        public decimal? COUNT { get; set; }
+        public int? COUNT { get; set; }
         public string STATUS { get; set; }
         public string DESCRIPTION { get; set; }
